Perform all transforms in double precision. Point coords are still in float.

+++ Objtree.h
@@ -297,8 +297,8 @@ typedef struct Transform
     float           ry;
     float           rz;
     BOOL            enable_rotation; // TRUE to allow the rotation to be applied
-    float           mat[9];         // Net matrix from scales and rotates
-    float           inv_mat[9];     // Inverse matrix
+    double          mat[9];         // Net matrix from scales and rotates
+    double          inv_mat[9];     // Inverse matrix
     TRANSFLAGS      flags;          // Transform flag bits, that allow optimisation of the speed of application.
 } Transform;
